+++ runtime/runtime_test.go
@@ -8195,8 +8195,7 @@ func TestRuntimeDestructorReentrancyPrevention(t *testing.T) {
 	)
 	RequireError(t, err)
 
-	var destroyedResourceErr interpreter.DestroyedResourceError
-	require.ErrorAs(t, err, &destroyedResourceErr)
+	require.ErrorAs(t, err, &interpreter.InvalidatedResourceReferenceError{})
 }
 
 func TestRuntimeFlowEventTypes(t *testing.T) {
@@ -8636,6 +8635,5 @@ func TestInvalidatedResourceUse2(t *testing.T) {
 
 	RequireError(t, err)
 
-	var destroyedResourceErr interpreter.DestroyedResourceError
-	require.ErrorAs(t, err, &destroyedResourceErr)
+	require.ErrorAs(t, err, &interpreter.InvalidatedResourceReferenceError{})
 }
